rag/linux_code_summary/requirements.txt
anthropic
fastapi
uvicorn[standard]
chromadb
sentence-transformers
tree-sitter
tree-sitter-c
python-dotenv
pydantic
streamlit
